- Solve bug language switch: if in the browser was set an unsupported language than the language switch get a NullPointerException. For this we have add an language default property in the ingrid-portal-apps.properties which set an default language
- Issue GSSOIL-7: (http://88.198.11.89/jira/browse/GSSOIL-7)
RSS feeds will not be read. In the past we have a parameter for read RSS news. Old RSS news will be delete always in a month. Now we use only one parameter which handle the reading and deletion of the RSS news, e.x. PortalU: Read RSS news which is younger than 30 days and delete RSS news which is older than 30 days. For GS Soil the parameter has the value 360 days.