Backup partner and provider table
With 4.4.0 the partner and provider are received from the codelist repository.
The previous partners and providers can be overwritten, so a backup is created
into the portal-directory. The created files are patch files, that can be used
in the "data/patches"-directory of the codelist-repository and the previous
values will be used in there.