Add new context help feature #272
New context help feature based on markdown files. The
affiliation to the GUI Element cat be set via metadata
in the markdown files.
If no drop in replacement in form of a  markdown file
can be found, the old database based help will be
used instead.

Further improvements include localization and profiling.